Adding a unit test to demonstrate and detect issue #116
Create a parent-child model where the child has a FK reference to
the parent. Have the parent have an IList collection of children.

EF will automatically put a HashSet in for ICollection, so it must
be IList to demonstrate the issue.

The test then uses OwnedCollection to demonstrate multiple adding
of children to the parent element.